Update _config.yml, removed parchment from baseurl
When I changed the URL in the configuration file, I accidentally broke all CSS formatting, causing the webpage to only be displayed as HTML. This happened because I only changed the URL, not the base URL, and it caused an error. Why? The forked version, which was copied straight from the creator, had the URL rahulbothra.com and the base URL "/parchment". This means that he had a domain outside Github pages as his website URL, and he was hosting his site in a subdirectory. Since I am not hosting my site in a subdirectory, I can leave the baseURL empty. So, while it was correct to change my URL to my specific Github Pages username, I neglected to delete "/parchment" so that my URL matched my base URL.